Add Result<()> return type to GammaMethod.set_temperature()
Allow gamma method to return return an error in the event of failure.

It appears that the Rust XCB bindings have trouble actually getting
randr to do anything, but perhaps the missing call to request_check()
after calling randr::set_crtc_gamma_checked() is the correct trigger. It
seems to get --oneshot to work, but --reset is still not responsive.