location: Move location provider into submodule
It should be possible to use geoclue2 for location determination, but I
have yet to figure out how to interact with dbus.

Adds a "geoclue2" feature to Cargo.toml and depends on dbus crate if
that configuration feature is given. It is not possible to use it for
requests yet.