Prevent crazy brightness values
By accident, I used -b instead of -t with the arguments 4500:3500. The
brightness is supposed to be between 0.1 and 1.0, but this was not checked, and
the brightness adjustment would use any value that parsed as a float.

Add a test and implement a check in parse_brightness() that prevents values
outside the allowed range to be used.